requirements.txt
mcp>=1.0.0
Pygments>=2.15.0
Pillow>=10.0.0
